+++ FileKit/Core/DataType.swift
@@ -49,7 +49,7 @@ extension Readable {
     ///
     /// - Parameter path: The path being read from.
     ///
-    public init(contentsOfPath path: Path) throws {
+    public init(contentsOfPath path: Path) throws { // swiftlint:disable:this valid_docs
         self = try Self.readFromPath(path)
     }
 
@@ -81,7 +81,7 @@ extension Writable {
     ///
     /// - Parameter path: The path being written to.
     ///
-    public func writeToPath(path: Path) throws {
+    public func writeToPath(path: Path) throws { // swiftlint:disable:this valid_docs
         try writeToPath(path, atomically: true)
     }
