+++ FileKit/Core/DirectoryEnumerator.swift
@@ -34,6 +34,8 @@ public struct DirectoryEnumerator: GeneratorType {
     private let _path: Path, _enumerator: NSDirectoryEnumerator?
 
     /// Creates a directory enumerator for the given path.
+    ///
+    /// - Parameter path: The path a directory enumerator to be created for.
     public init(path: Path) {
         _path = path
         _enumerator = NSFileManager().enumeratorAtPath(path.rawValue)
