+++ FileKit/Core/File.swift
@@ -66,6 +66,8 @@ public class File<Data: DataType>: Comparable {
     // MARK: - Initialization
 
     /// Initializes a file from a path.
+    ///
+    /// - Parameter path: The path a file to initialize from.
     public init(path: Path) {
         self.path = path
     }
@@ -75,6 +77,7 @@ public class File<Data: DataType>: Comparable {
     /// Reads the file and returns its data.
     ///
     /// - Throws: `FileKitError.ReadFromFileFail`
+    /// - Returns: The data read from file.
     public func read() throws -> Data {
         return try Data.readFromPath(path)
     }
@@ -116,6 +119,11 @@ public class File<Data: DataType>: Comparable {
     }
 
     /// Deletes the file.
+    ///
+    /// Throws an error if the file could not be deleted.
+    ///
+    /// - Throws: `FileKitError.DeleteFileFail`
+    ///
     public func delete() throws {
         try path.deleteFile()
     }
@@ -126,6 +134,7 @@ public class File<Data: DataType>: Comparable {
     ///
     /// Throws an error if the file cannot be moved.
     ///
+    /// - Parameter path: The path to move the file to.
     /// - Throws: `FileKitError.MoveFileFail`
     ///
     public func moveToPath(path: Path) throws {
@@ -138,6 +147,8 @@ public class File<Data: DataType>: Comparable {
     /// Throws an error if the file could not be copied or if a file already
     /// exists at the destination path.
     ///
+    ///
+    /// - Parameter path: The path to copy the file to.
     /// - Throws: `FileKitError.FileDoesNotExist`, `FileKitError.CopyFileFail`
     ///
     public func copyToPath(path: Path) throws {
@@ -152,6 +163,8 @@ public class File<Data: DataType>: Comparable {
     /// If the path already exists and _is_ a directory, the link will be made
     /// to `self` in that directory.
     ///
+    ///
+    /// - Parameter path: The path to symlink the file to.
     /// - Throws:
     ///     `FileKitError.FileDoesNotExist`,
     ///     `FileKitError.CreateSymlinkFail`
