+++ FileKit/Core/Operators.swift
@@ -25,6 +25,8 @@
 //  THE SOFTWARE.
 //
 
+// swiftlint:disable file_length
+
 import Foundation
 
 // MARK: - File
