+++ FileKit/Core/Path.swift
@@ -507,6 +507,8 @@ extension Path {
     // MARK: - RawRepresentable
 
     /// Initializes a path to the string value.
+    ///
+    /// - Parameter rawValue: The raw value to initialize from.
     public init(rawValue: String) {
         self.rawValue = rawValue
     }
@@ -685,6 +687,9 @@ extension Path {
 
     // MARK: - NSURL
 
+    /// Creates a new path with given url if possible.
+    ///
+    /// - Parameter url: The url to create a path for.
     public init?(url: NSURL) {
         guard let path = url.path where url.fileURL else {
             return nil
@@ -692,6 +697,7 @@ extension Path {
         rawValue = path
     }
 
+    /// - Returns: The `Path` objects url.
     public var url: NSURL {
         return NSURL(fileURLWithPath: rawValue, isDirectory: self.isDirectory)
     }
@@ -702,6 +708,9 @@ extension Path {
 
     // MARK: - BookmarkData
 
+    /// Creates a new path with given url if possible.
+    ///
+    /// - Parameter bookmarkData: The bookmark data to create a path for.
     public init?(bookmarkData bookData: NSData) {
         var isStale: ObjCBool = false
         let url = try? NSURL(
@@ -715,6 +724,7 @@ extension Path {
         self.init(url: fullURL)
     }
 
+    /// - Returns: The `Path` objects bookmarkData.
     public var bookmarkData: NSData? {
         return try? url.bookmarkDataWithOptions(
             .SuitableForBookmarkFile,
@@ -729,6 +739,8 @@ extension Path {
     // MARK: - SecurityApplicationGroupIdentifier
 
     /// Returns the container directory associated with the specified security application group ID.
+    ///
+    /// - Parameter groupIdentifier: The group identifier.
     public init?(groupIdentifier: String) {
         guard let url = NSFileManager().containerURLForSecurityApplicationGroupIdentifier(groupIdentifier) else {
             return nil
@@ -832,7 +844,7 @@ extension Path: SequenceType {
 
     // MARK: - SequenceType
 
-    /// Return a *generator* over the contents of the path.
+    /// - Returns: A *generator* over the contents of the path.
     public func generate() -> DirectoryEnumerator {
         return DirectoryEnumerator(path: self)
     }
